v0.4.4//bugfix:navigation

+++ frontend/src/pages/RegisterPage/RegisterPage.jsx
@@ -52,15 +52,16 @@ export default function RegisterPage() {
     });
   }, [values]);
   useEffect(() => {
-    // confirmData && console.log('data:', confirmData);
-    confirmData && localStorage.setItem('token', confirmData.token);
-    confirmData && dispatch(setUserData({
-      username: confirmData.username,
-      email: confirmData.email,
-      isAuth: true
-    }));
-    (confirmData && location.state?.from) ?
-      navigate(location.state.from) : navigate(-1);
+    if (confirmData) {
+      confirmData && console.log('1');
+      confirmData && localStorage.setItem('token', confirmData.token);
+      confirmData && dispatch(setUserData({
+        username: confirmData.username,
+        email: confirmData.email,
+        isAuth: true
+      }));
+      location.state?.from ? navigate(location.state.from) : navigate(-1);
+    }
     confirmError && console.log('error status:', confirmError.status);
     confirmError && console.log('error data:', confirmError.data);
   }, [confirmData, confirmIsSuccess, confirmIsError]);
